backend/requirements/base.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
django-filter==23.3
drf-spectacular==0.26.5

# MongoDB stack
pymongo==4.5.0
mongoengine==0.27.0

# PDF/Doc parsing
PyPDF2==3.0.1
pdfplumber==0.9.0
python-docx==1.1.0
python-magic==0.4.27

# GenAI
langchain==0.0.340
openai==1.3.7
tiktoken==0.5.2
transformers==4.35.2

# Testing
pytest==8.4.1
pytest-django==4.11.1
pytest-cov==6.2.1

# HTTP requests
requests==2.31.0
